add transaction id for commission fee

+++ CRM.DAL/Models/CommissionFeeDto.cs
@@ -8,7 +8,7 @@ namespace CRM.DAL.Models
         public int Id { get; set; }
         public int LeadId { get; set; }
         public int AccountId { get; set; }
-        public int TransactionId { get; set; }
+        public long TransactionId { get; set; }
         public Role Role { get; set; }
         public DateTime Date { get; set; }
         public decimal Amount { get; set; }
